cambios en aplicativo

+++ lib/screens/brand_screen.dart
@@ -1,3 +1,4 @@
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:flutter/cupertino.dart';
@@ -144,6 +145,22 @@ class _BrandScreenState extends State<BrandScreen> {
       _showLoader = true;
     });
     
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
+    
     Map<String, dynamic> request ={
       'id': widget.brand.id,
       'description': _description,
@@ -152,7 +169,7 @@ class _BrandScreenState extends State<BrandScreen> {
     Response response = await ApiHelper.post(
       '/api/Brands/', 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -178,6 +195,22 @@ class _BrandScreenState extends State<BrandScreen> {
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Map<String, dynamic> request ={
       'id': widget.brand.id,
@@ -188,7 +221,7 @@ class _BrandScreenState extends State<BrandScreen> {
       '/api/Brands/', 
       widget.brand.id.toString(), 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -231,11 +264,27 @@ class _BrandScreenState extends State<BrandScreen> {
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Response response = await ApiHelper.delete(
       '/api/Brands/', 
       widget.brand.id.toString(), 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
