+++ lib/screens/documenttype_screen.dart
@@ -1,3 +1,4 @@
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:flutter/cupertino.dart';
@@ -144,6 +145,22 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Map<String, dynamic> request ={
       'id': widget.documenttype.id,
@@ -153,7 +170,7 @@ bool _showLoader = false;
     Response response = await ApiHelper.post(
       '/api/DocumentTypes/', 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -179,6 +196,22 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Map<String, dynamic> request ={
       'id': widget.documenttype.id,
@@ -189,7 +222,7 @@ bool _showLoader = false;
       '/api/DocumentTypes/', 
       widget.documenttype.id.toString(), 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -232,11 +265,27 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Response response = await ApiHelper.delete(
       '/api/DocumentTypes/', 
       widget.documenttype.id.toString(), 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
