+++ lib/screens/vehicletype_screen.dart
@@ -1,3 +1,4 @@
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 import 'package:adaptive_dialog/adaptive_dialog.dart';
 import 'package:flutter/cupertino.dart';
@@ -145,6 +146,22 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Map<String, dynamic> request ={
       'id': widget.vehicletype.id,
@@ -154,7 +171,7 @@ bool _showLoader = false;
     Response response = await ApiHelper.post(
       '/api/VehicleTypes/', 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -180,6 +197,22 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Map<String, dynamic> request ={
       'id': widget.vehicletype.id,
@@ -190,7 +223,7 @@ bool _showLoader = false;
       '/api/VehicleTypes/', 
       widget.vehicletype.id.toString(), 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -233,11 +266,27 @@ bool _showLoader = false;
     setState(() {
       _showLoader = true;
     });
+
+    var connectivityResult = await Connectivity().checkConnectivity();
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+      await showAlertDialog(
+        context: context,
+        title: 'Error', 
+        message: 'Verifica que estes conectado a internet.',
+        actions: <AlertDialogAction>[
+            AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );    
+      return;
+    }
     
     Response response = await ApiHelper.delete(
       '/api/VehicleTypes/', 
       widget.vehicletype.id.toString(), 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
